WIP5: Checkout Logic
