+++ src/common/Loader/index.js
@@ -6,14 +6,14 @@ import styles from './index.module.scss';
 
 import LoaderImage from 'assets/images/loader.png';
 
-const Loader = ({ noPortal }) => {
+const Loader = ({ className, noPortal }) => {
   const backdropElement = document.getElementById('backdrop');
   const overlaysElement = document.getElementById('overlays');
 
   if (noPortal) {
     return (
       <>
-        <div className={styles.loader_wrapper}>
+        <div className={`${className} ${styles.loader_wrapper}`}>
           <img className={styles.loader} src={LoaderImage} alt="" />
         </div>
       </>
