+++ src/pages/Checkout/CheckoutProgression/index.js
@@ -2,7 +2,7 @@ import Step from './Step';
 
 import styles from './index.module.scss';
 
-const CheckoutProgression = ({ handleSelectStep, steps, currentStep }) => {
+const CheckoutProgression = ({ steps }) => {
   return (
     <div className={styles.steps_container}>
       {steps.map((step, index) => (
@@ -11,8 +11,6 @@ const CheckoutProgression = ({ handleSelectStep, steps, currentStep }) => {
           label={step.label}
           url={step.url}
           index={index}
-          handleSelectStep={handleSelectStep}
-          currentStep={currentStep}
         />
       ))}
     </div>
