+++ src/pages/Checkout/ShippingInfo/index.module.scss
@@ -37,6 +37,11 @@
   }
 }
 
+.loader_wrapper {
+  height: auto!important;
+  padding-top: 4rem;
+}
+
 .info_form {
   margin-top: 4rem;
   display: flex;
