+++ src/pages/Checkout/ShippingOption/index.module.scss
@@ -1,4 +1,4 @@
-.shipping_wrapper {
+.shipping_option_wrapper {
   h2 {
     color: #fff;
     margin-bottom: 1.5rem;
@@ -14,7 +14,7 @@
   }
 }
 
-.shipping_form {
+.shipping_option_form {
   color: #fff;
   border: 1px solid #515255;
   padding: 1rem 2rem;
